Update to minimum Java 7
