+++ src/test/antunit/Java7ClassReferences.java
@@ -17,26 +17,22 @@
 /* The binary class file is packaged together with the source distribution.
  */
 
-import java.lang.reflect.Field;
-import java.lang.reflect.Method;
 import java.util.*;
 
-class Java6Reflection {
-  static Field test() throws Exception {
-    Class c = Java6Reflection.class;
-    Field f = c.getDeclaredField("field1");
-    f.setAccessible(true);
-    
-    Method m = c.getDeclaredMethod("testMethod");
-    m.setAccessible(true);
-    m.invoke(new Java6Reflection());
-    
-    return f;
+class Java7ClassReferences {
+  static Integer[][] test() {
+    Integer.class.getName();
+    System.out.println(Integer[].class);
+    Integer[] arr = new Integer[1];
+    arr[0] = new Integer(0);
+    return new Integer[1][1];
   }
   
   private Integer field1;
-
-  private void testMethod() {
-    // nothing to do here
-  }
+  private final Integer[] field2 = null;
+  
+  // we forbid the superclass or interface, but concrete instance is used:
+  public List<?> list1;
+  public ArrayList<?> list2;
+  public HashSet<?> set1;
 }
